requirements.txt
asgiref==3.7.2
attrs==23.2.0
dj-rest-auth==5.0.2
Django==4.2.9
django-cleanup==8.0.0
django-cors-headers==4.3.1
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
phonenumberslite==8.13.25
pillow==10.2.0
psycopg2-binary==2.9.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
rpds-py==0.17.1
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
